+++ djwebhooks/senders/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from .orm import orm_callable
+from .redislog import redislog_callable
 from .redisq import redisq_callable